Create temporary directory to pack project under target/
The temporary directory used to pack the project is created under the
current directory. For a multimodules project, if modules are built
concurrently, it means there's an opportunity for 2 modules tries to
create the same temporary directory with one of them failing.

To prevent this issue, create the temporary directory under the build
output directory (which also make cleaning easier if process crashes).